requirements.txt
# Core ML & Data Handling
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.0.3
imbalanced-learn==0.12.3
pyarrow==16.1.0

# Visualization (dev)
matplotlib==3.9.0
seaborn==0.13.2

# API & Server
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
streamlit==1.35.0
PyYAML==6.0.1

# Utils
python-dotenv==1.0.1
joblib==1.4.2
